Re-introduce executable functionality

+++ samplePDFDUNE/samplePDFDUNEBeamFD.cpp
@@ -378,7 +378,7 @@ int samplePDFDUNEBeamFD::setupExperimentMC(int iSample) {
     duneobj->Target[i] = 40;
     
     int mode= TMath::Abs(_mode);       
-    duneobj->mode[i]=SIMBMode_ToMaCh3Mode(mode, _isCC);
+    duneobj->mode[i]=(double)SIMBMode_ToMaCh3Mode(mode, _isCC);
     
     duneobj->flux_w[i] = 1.0;
   }
